+++ now/constants.py
@@ -7,7 +7,7 @@ from now.utils import BetterEnum
 # DEMO_DATASET_DOCARRAY_VERSION = docarray_version
 DEMO_DATASET_DOCARRAY_VERSION = '0.13.17'
 
-DOCKER_BFF_PLAYGROUND_TAG = '0.0.129-refactor-bff-2'
+DOCKER_BFF_PLAYGROUND_TAG = '0.0.129-refactor-bff-3'
 NOW_PREPROCESSOR_VERSION = '0.0.94-allow-domain-6'
 NOW_QDRANT_INDEXER_VERSION = '0.0.2-allow-domain-6'
 NOW_ELASTIC_INDEXER_VERSION = '0.0.3-refactor-executors-21'
